https://numworx.atlassian.net/browse/DWOPLAY-198
bij diff geen cas maar zelf testen met steekproef

git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DwoPlayer/GWT/trunk@27438 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/fi/wiskopdr/AntwoordVergelijkingVakChecker.java
@@ -501,7 +501,7 @@ public class AntwoordVergelijkingVakChecker implements AntwoordVakChecker
 		if(answerModels==null || answerModels.get(answerModelNr)==null)
 			return 0;
 		ObjectMap map = JSONUtilities.wrapMap(answerModels.get(answerModelNr));
-		return (Integer)map.getInt("puntenFeedback");
+		return map.getInt("puntenFeedback");
 	}
